Added script to monitor WIFI network
This is script is based on OpenWRT's watchcat idea. The key principel is that if the camera cannot connec to a WIFI it is worthless so it willl try by all means to connect. Depending on the configuration, it will first try to restart the network or connec to to alternative SSIDs. If all this fails, after a couple of tries, it will reboot the camera.